fix: API 호출 로직 생성
