+++ .env
@@ -1,2 +1,3 @@
 REACT_APP_API_KEY=81a92362658e35f04c2b390845ea5f68
 REACT_APP_TMDB_BASE_URL=https://api.themoviedb.org/3/movie
+REACT_APP_TMDB_POSTER_URL=https://media.themoviedb.org/t/p
