+++ src/common/MovieCard/MovieCard.tsx
@@ -1,3 +1,4 @@
+import { fetchReleaseDate } from '@/apis/tmdb';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
@@ -9,10 +10,10 @@ interface Movie {
     year: number;
 }
 
-interface ReleaseDatesResponse {
-    iso_3166_1: string; //국가: 한국
-    release_dates: { release_date: string }[];
-}
+// interface ReleaseDatesResponse {
+//     iso_3166_1: string; //국가: 한국
+//     release_dates: { release_date: string }[];
+// }
 
 const MovieCard = ({ movie }: { movie: Movie }) => {
     const [isHovered, setIsHovered] = useState(false);
@@ -24,39 +25,39 @@ const MovieCard = ({ movie }: { movie: Movie }) => {
     };
 
     // TMDb API에서 영화의 개봉일 정보 가져오는 함수
-    const fetchReleaseDate = async (movieId: number) => {
-        const apiKey = process.env.REACT_APP_API_KEY;
-        const movieUrl = process.env.REACT_APP_TMDB_BASE_URL;
-        // const url = `https://api.themoviedb.org/3/movie/${movieId}/release_dates?api_key=${apiKey}`;
-        const url = `${movieUrl}/${movieId}/release_dates?api_key=${apiKey}`;
-        try {
-            const response = await fetch(url);
-            if (!response.ok) {
-                console.error('Failed to fetch release dates');
-                setReleaseDate(null); // 실패 시 명시적으로 null 설정
-                return;
-            }
-            const data = await response.json();
-            const releaseDates: ReleaseDatesResponse[] = data.results;
+    // const fetchReleaseDate = async (movieId: number) => {
+    //     const apiKey = process.env.REACT_APP_API_KEY;
+    //     const movieUrl = process.env.REACT_APP_TMDB_BASE_URL;
+    //     // const url = `https://api.themoviedb.org/3/movie/${movieId}/release_dates?api_key=${apiKey}`;
+    //     const url = `${movieUrl}/${movieId}/release_dates?api_key=${apiKey}`;
+    //     try {
+    //         const response = await fetch(url);
+    //         if (!response.ok) {
+    //             console.error('Failed to fetch release dates');
+    //             setReleaseDate(null); // 실패 시 명시적으로 null 설정
+    //             return;
+    //         }
+    //         const data = await response.json();
+    //         const releaseDates: ReleaseDatesResponse[] = data.results;
 
-            const koreanRelease = releaseDates.find(
-                release => release.iso_3166_1 === 'KR',
-            );
+    //         const koreanRelease = releaseDates.find(
+    //             release => release.iso_3166_1 === 'KR',
+    //         );
 
-            if (
-                koreanRelease &&
-                koreanRelease.release_dates &&
-                koreanRelease.release_dates.length > 0
-            ) {
-                setReleaseDate(koreanRelease.release_dates[0].release_date);
-            } else {
-                setReleaseDate(null); // 적합한 데이터가 없을 경우
-            }
-        } catch (error) {
-            console.error('Error fetching release date:', error);
-            setReleaseDate(null); // 에러 발생 시 null 설정
-        }
-    };
+    //         if (
+    //             koreanRelease &&
+    //             koreanRelease.release_dates &&
+    //             koreanRelease.release_dates.length > 0
+    //         ) {
+    //             setReleaseDate(koreanRelease.release_dates[0].release_date);
+    //         } else {
+    //             setReleaseDate(null); // 적합한 데이터가 없을 경우
+    //         }
+    //     } catch (error) {
+    //         console.error('Error fetching release date:', error);
+    //         setReleaseDate(null); // 에러 발생 시 null 설정
+    //     }
+    // };
 
     useEffect(() => {
         if (!movie.id) return; // movie.id가 없으면 fetch 호출하지 않음
