+++ webpack.config.mjs
@@ -54,6 +54,9 @@ const config = {
             'process.env.REACT_APP_TMDB_BASE_URL': JSON.stringify(
                 process.env.REACT_APP_TMDB_BASE_URL,
             ),
+            'process.env.REACT_APP_TMDB_POSTER_URL': JSON.stringify(
+                process.env.REACT_APP_TMDB_POSTER_URL,
+            ),
         }),
     ],
     module: {
